Added plain-text version of the main AppleScript to the bundle.
The plain-text version is purely for reference, it has no effect on the app's operation. And it has to be manually exported each time main.scpt is changed, so maybe if that becomes a hassle I'll remove it. For now I'm thinking it'll be useful in order to see diffs as changes are made to main.scpt.